The 86th koan passes.
Even though it would be sort of a useless result, it seems like /z*/
should have matched the entire string for consistency. I need to go
revisit my automata theory, I suppose. Until then, I can only guess that
the kleene star (I just impressed myself by remembering the actual name
for that thing) is defined not to match in this case precisely because
it's a useless result.