The 133rd koan passes.
It's nice that the if/else constructs have return values, I guess, but
I'm glad this little C-ism is still around. It's much more concise, and
to me it says "Hey, we care about the return value here", whereas the
normal behavior is to not pay much attention to the return value of
if/else statements.